feat: Build web UI for showing streaming line chart and historical line chart by using plotly
When user click on live channel, show the loading message until received data for ploting the chart.
Show the average viewer count of the chatroom of selected channel.

when user click on default channel, also show the loading message.
show the selected channel name and historical live schedule which are able to be clicked.
Show the historical line chart of selected channel and the date of 'startedAt'